revert to fixture instead of mixin for tests
